refactor(tests): convert coverage-padding tests to behavioral tests
Story 7.1: Refactor all 16 coverage-named test files into well-organized
behavioral tests with strong assertions.

Changes:
- Delete 11 coverage-padding test files (*_coverage.py, *_extra.py)
- Rename 5 files to behavioral names (e.g., test_logger_initialization.py)
- Add 4 new consolidated behavioral test files:
  - test_error_classes.py (error serialization)
  - test_logger_error_containment.py (exception handling)
  - test_logger_features.py (enrichers, sampling, metrics)
  - test_logger_thread_lifecycle.py (thread mode, backpressure)
- Fix all weak assertions (>= 0, >= 1) on result counters
- Remove line-number references from docstrings
- Maintain 90%+ coverage (90.47%)

Net result: -3,400 lines, same coverage, stronger assertions.
