--- tests/unit/test_core_logger_exceptional_coverage.py
@@ -1,868 +1,0 @@
-"""
-Exceptional coverage tests for src/fapilog/core/logger.py
-
-This module contains targeted tests to achieve 90%+ coverage by testing
-exception paths, edge cases, and error handling scenarios that are critical
-for a production logging library.
-"""
-
-import asyncio
-import threading
-import time
-from typing import Any
-from unittest.mock import Mock, patch
-
-import pytest
-
-from fapilog.core.events import LogEvent
-from fapilog.core.logger import AsyncLogger, AsyncLoggerFacade, SyncLoggerFacade
-from fapilog.metrics.metrics import MetricsCollector
-
-
-def _create_async_sink(collected: list[dict[str, Any]]):
-    """Create an async sink function."""
-
-    async def async_sink(event: dict[str, Any]) -> None:
-        collected.append(dict(event))
-
-    return async_sink
-
-
-class TestAsyncLoggerPlaceholder:
-    """Test the minimal AsyncLogger placeholder class (line 34)."""
-
-    @pytest.mark.asyncio
-    async def test_async_logger_log_many(self) -> None:
-        """Test AsyncLogger.log_many placeholder method."""
-        logger = AsyncLogger()
-
-        # Test with empty events
-        result = await logger.log_many([])
-        assert result == 0
-
-        # Test with multiple events
-        events = [
-            LogEvent(message="test1"),
-            LogEvent(message="test2"),
-            LogEvent(message="test3"),
-        ]
-        result = await logger.log_many(events)
-        assert result == 3
-
-
-class TestExceptionPathsInEnqueue:
-    """Test exception handling paths in enqueue methods."""
-
-    def test_sync_enqueue_exception_handling_lines_413_428(self) -> None:
-        """Test exception handling in sync enqueue (lines 413-428)."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="exception-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        # Force an exception in the enqueue path by mocking queue operations
-        with patch.object(
-            logger._queue, "try_enqueue", side_effect=RuntimeError("Queue error")
-        ):
-            # This should trigger the exception handling path (lines 415-428)
-            logger.info("test message")
-
-            # The message should be marked as dropped
-            assert logger._dropped >= 1
-
-    @pytest.mark.asyncio
-    async def test_async_enqueue_exception_handling(self) -> None:
-        """Test exception handling in async enqueue methods."""
-        out: list[dict[str, Any]] = []
-
-        logger = AsyncLoggerFacade(
-            name="async-exception-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=_create_async_sink(out),
-        )
-
-        logger.start()
-
-        # Force an exception in async enqueue
-        with patch.object(
-            logger._queue, "try_enqueue", side_effect=RuntimeError("Async queue error")
-        ):
-            try:
-                await logger.info("test message")
-            except RuntimeError:
-                # Exception is expected to propagate in this case
-                pass
-
-            # The exception should cause the enqueue to fail
-            # This tests the exception path but doesn't guarantee drop count
-            assert True  # Just verify no crash
-
-        await logger.stop_and_drain()
-
-
-class TestSamplingExceptionPaths:
-    """Test sampling logic exception handling (lines 996-1001)."""
-
-    def test_sampling_with_settings_exception(self) -> None:
-        """Test sampling when Settings() raises an exception."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="sampling-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock Settings to raise an exception
-        with patch(
-            "fapilog.core.settings.Settings", side_effect=RuntimeError("Settings error")
-        ):
-            # Should handle the exception and continue (line 1000-1001)
-            logger.debug("Debug message that should still be processed")
-            logger.info("Info message that should still be processed")
-
-        asyncio.run(logger.stop_and_drain())
-
-        # Messages should still be processed despite settings exception
-        # At least one message should get through
-        assert len(out) >= 1
-        # Note: processed count might be 0 if messages are dropped due to exceptions
-        assert len(out) >= 1
-
-    def test_sampling_with_random_exception(self) -> None:
-        """Test sampling when random.random() raises an exception."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="sampling-random-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock random.random to raise an exception
-        with patch("random.random", side_effect=RuntimeError("Random error")):
-            # Should handle the exception and continue
-            logger.debug("Debug message")
-            logger.info("Info message")
-
-        asyncio.run(logger.stop_and_drain())
-
-        # Messages should still be processed despite random exception
-        assert len(out) >= 1
-
-    def test_sampling_boundary_conditions(self) -> None:
-        """Test sampling at boundary conditions."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="sampling-boundary-test",
-            queue_capacity=32,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Test with sampling rate exactly at boundary
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            settings_instance = Mock()
-            settings_instance.core.log_sampling_rate = 0.0  # 0% sampling
-            mock_settings.return_value = settings_instance
-
-            # Mock random to return values at boundaries
-            with patch("random.random", side_effect=[0.0, 0.5, 1.0]):
-                logger.debug("Debug 1")  # Should be filtered (0.0 > 0.0 is False)
-                logger.debug("Debug 2")  # Should be filtered (0.5 > 0.0 is True)
-                logger.debug("Debug 3")  # Should be filtered (1.0 > 0.0 is True)
-
-        asyncio.run(logger.stop_and_drain())
-
-        # With 0% sampling, very few DEBUG messages should get through
-        # But due to exception handling, some might still pass
-        # Just verify we got some messages and no crash
-        assert len(out) >= 0
-
-
-class TestErrorDeduplicationEdgeCases:
-    """Test error deduplication edge cases and exception paths."""
-
-    def test_error_dedup_with_settings_exception(self) -> None:
-        """Test error deduplication when Settings() raises exception."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="dedup-settings-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock Settings to raise exception on second call (inside dedup logic)
-        settings_calls = 0
-
-        def settings_side_effect():
-            nonlocal settings_calls
-            settings_calls += 1
-            if settings_calls == 2:  # Second call in dedup logic
-                raise RuntimeError("Settings error in dedup")
-            settings_instance = Mock()
-            settings_instance.core.error_dedupe_window_seconds = 5.0
-            return settings_instance
-
-        with patch("fapilog.core.settings.Settings", side_effect=settings_side_effect):
-            # First error should work normally
-            logger.error("Database error")
-            # Second error should hit the settings exception but continue
-            logger.error("Database error")
-
-        asyncio.run(logger.stop_and_drain())
-        # Should handle the exception gracefully and still process messages
-        assert len(out) >= 1
-
-    def test_error_dedup_with_time_monotonic_exception(self) -> None:
-        """Test error deduplication when time.monotonic() raises exception."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="dedup-time-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Test that error deduplication works normally
-        logger.error("Error message 1")
-        logger.error("Error message 1")  # Duplicate - should be suppressed
-        logger.error("Error message 2")  # Different - should be logged
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle deduplication gracefully; duplicate may be suppressed
-        assert len(out) >= 1  # At least one message should be processed
-
-    def test_error_dedup_with_settings_exception_duplicate(self) -> None:
-        """Test error deduplication when Settings() raises exception."""
-        out: list[dict[str, Any]] = []
-
-        # Mock Settings to raise exception during deduplication
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            mock_settings.side_effect = RuntimeError("Settings failed")
-
-            logger = SyncLoggerFacade(
-                name="dedup-settings-exception-test",
-                queue_capacity=16,
-                batch_max_size=4,
-                batch_timeout_seconds=0.01,
-                backpressure_wait_ms=1,
-                drop_on_full=False,
-                sink_write=lambda e: out.append(e),
-            )
-
-            logger.start()
-
-            # Should handle Settings exception gracefully
-            logger.error("Error message with settings exception")
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle the exception gracefully
-            assert len(out) >= 1  # Message should still be processed
-            assert len(out) >= 1
-
-    def test_sampling_with_settings_exception(self) -> None:
-        """Test sampling when Settings() raises exception."""
-        out: list[dict[str, Any]] = []
-
-        # Mock Settings to raise exception during sampling
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            mock_settings.side_effect = RuntimeError("Settings failed")
-
-            logger = SyncLoggerFacade(
-                name="sampling-settings-exception-test",
-                queue_capacity=16,
-                batch_max_size=4,
-                batch_timeout_seconds=0.01,
-                backpressure_wait_ms=1,
-                drop_on_full=False,
-                sink_write=lambda e: out.append(e),
-            )
-
-            logger.start()
-
-            # Should handle Settings exception gracefully
-            logger.info("Info message with settings exception")
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle the exception gracefully
-            assert len(out) >= 1  # Message should still be processed
-            assert len(out) >= 1
-
-    def test_sampling_logic_with_low_rate(self) -> None:
-        """Test sampling logic when rate < 1.0."""
-        out: list[dict[str, Any]] = []
-
-        # Mock Settings to return a low sampling rate
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            mock_instance = Mock()
-            mock_instance.observability.logging.sampling_rate = 0.1  # 10% sampling
-            mock_settings.return_value = mock_instance
-
-            logger = SyncLoggerFacade(
-                name="sampling-low-rate-test",
-                queue_capacity=16,
-                batch_max_size=4,
-                batch_timeout_seconds=0.01,
-                backpressure_wait_ms=1,
-                drop_on_full=False,
-                sink_write=lambda e: out.append(e),
-            )
-
-            logger.start()
-
-            # Send multiple DEBUG messages - some should be sampled out
-            for i in range(10):
-                logger.debug(f"Debug message {i}")
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle sampling gracefully
-            assert len(out) >= 0  # Some messages may be sampled out
-            assert len(out) >= 0
-
-    def test_context_variable_exception_handling(self) -> None:
-        """Test context variable exception handling."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="context-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock the bound context variable to raise exception
-        with patch.object(logger, "_bound_context_var") as mock_context:
-            mock_context.get.side_effect = RuntimeError("Context failed")
-
-            # Should handle context exception gracefully
-            logger.info("Info message with context exception")
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle the exception gracefully
-            assert len(out) >= 1  # Message should still be processed
-            assert len(out) >= 1
-
-    def test_exception_serialization_exception_handling(self) -> None:
-        """Test exception serialization exception handling."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="exception-serialization-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock serialize_exception to raise exception
-        with patch("fapilog.core.errors.serialize_exception") as mock_serialize:
-            mock_serialize.side_effect = RuntimeError("Serialization failed")
-
-            # Should handle serialization exception gracefully
-            logger.error("Error with exception", exc=RuntimeError("Test error"))
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle the exception gracefully
-            assert len(out) >= 1  # Message should still be processed
-            assert len(out) >= 1
-
-    def test_exception_serialization_with_exc_info_true(self) -> None:
-        """Test exception serialization with exc_info=True."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="exc-info-true-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock sys.exc_info to return a valid exception tuple
-        with patch("sys.exc_info") as mock_exc_info:
-            mock_exc_info.return_value = (RuntimeError, RuntimeError("Test"), None)
-
-            # Should handle exc_info=True gracefully
-            logger.error("Error with exc_info=True", exc_info=True)
-
-            # Allow processing time
-            time.sleep(0.05)
-
-            # Test shutdown
-            asyncio.run(logger.stop_and_drain())
-
-            # Should handle the exception gracefully
-            assert len(out) >= 1  # Message should still be processed
-            assert len(out) >= 1
-
-    def test_exception_serialization_with_exc_info_tuple(self) -> None:
-        """Test exception serialization with exc_info tuple."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="exc-info-tuple-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Test with exc_info tuple
-        exc_info = (RuntimeError, RuntimeError("Test tuple"), None)
-        logger.error("Error with exc_info tuple", exc_info=exc_info)
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle the exception gracefully
-        assert len(out) >= 1  # Message should still be processed
-        assert len(out) >= 1
-
-
-class TestMetricsExceptionPaths:
-    """Test metrics integration exception handling."""
-
-    def test_metrics_submission_exception_sync(self) -> None:
-        """Test metrics submission exception handling in sync logger."""
-        out: list[dict[str, Any]] = []
-
-        # Create a mock metrics collector that raises exceptions
-        mock_metrics = Mock(spec=MetricsCollector)
-        mock_metrics.record_events_submitted.side_effect = RuntimeError("Metrics error")
-
-        logger = SyncLoggerFacade(
-            name="metrics-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-            metrics=mock_metrics,
-        )
-
-        logger.start()
-
-        # Should handle metrics exceptions gracefully (lines 358-359)
-        logger.info("Test message with metrics exception")
-
-        asyncio.run(logger.stop_and_drain())
-
-        # Message should still be processed despite metrics exception
-        # But metrics errors might affect processing
-        assert len(out) >= 0  # At least verify no crash
-
-    @pytest.mark.asyncio
-    async def test_metrics_submission_exception_async(self) -> None:
-        """Test metrics submission exception handling in async logger."""
-        out: list[dict[str, Any]] = []
-
-        # Create a mock metrics collector that raises exceptions
-        mock_metrics = Mock(spec=MetricsCollector)
-        mock_metrics.record_events_submitted = Mock(
-            side_effect=RuntimeError("Async metrics error")
-        )
-
-        logger = AsyncLoggerFacade(
-            name="async-metrics-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=_create_async_sink(out),
-            metrics=mock_metrics,
-        )
-
-        logger.start()
-
-        # Should handle async metrics exceptions gracefully
-        await logger.info("Test message with async metrics exception")
-
-        await logger.stop_and_drain()
-
-        # Message should still be processed despite metrics exception
-        assert len(out) >= 1
-
-    def test_metrics_event_loop_exception_handling(self) -> None:
-        """Test metrics submission when no event loop is running."""
-        out: list[dict[str, Any]] = []
-
-        # Create a simple mock that doesn't need futures
-        mock_metrics = Mock(spec=MetricsCollector)
-        mock_metrics.record_events_submitted.return_value = None
-
-        logger = SyncLoggerFacade(
-            name="metrics-no-loop-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-            metrics=mock_metrics,
-        )
-
-        logger.start()
-
-        # Test normal operation first
-        logger.info("Test message in thread mode")
-
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle the operation gracefully
-        assert len(out) >= 1
-
-
-class TestSerializationFallbackPaths:
-    """Test serialization fast-path failures and fallback scenarios."""
-
-    @pytest.mark.asyncio
-    async def test_serialization_envelope_exception(self) -> None:
-        """Test serialization when envelope serialization fails."""
-        out: list[dict[str, Any]] = []
-        serialized_out: list[Any] = []
-
-        async def regular_sink(event: dict[str, Any]) -> None:
-            out.append(dict(event))
-
-        async def serialized_sink(view: Any) -> None:
-            serialized_out.append(view)
-
-        logger = SyncLoggerFacade(
-            name="serialization-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=regular_sink,
-            sink_write_serialized=serialized_sink,
-            serialize_in_flush=True,  # Enable serialization fast-path
-        )
-
-        logger.start()
-
-        # Mock serialize_envelope to raise an exception
-        with patch(
-            "fapilog.core.worker.serialize_envelope",
-            side_effect=RuntimeError("Serialization error"),
-        ):
-            # Should fall back to regular sink path
-            logger.info("Test message with serialization failure")
-
-        await logger.stop_and_drain()
-
-        # Should fall back or drop gracefully; just ensure no crash
-        assert len(out) >= 0
-
-    @pytest.mark.asyncio
-    async def test_serialization_strict_mode_exception(self) -> None:
-        """Test serialization in strict mode with exceptions."""
-        out: list[dict[str, Any]] = []
-        serialized_out: list[Any] = []
-
-        async def regular_sink(event: dict[str, Any]) -> None:
-            out.append(dict(event))
-
-        async def serialized_sink(view: Any) -> None:
-            serialized_out.append(view)
-
-        logger = SyncLoggerFacade(
-            name="serialization-strict-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=regular_sink,
-            sink_write_serialized=serialized_sink,
-            serialize_in_flush=True,
-        )
-
-        logger.start()
-
-        # Mock settings to enable strict mode and serialize_envelope to fail
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            settings_instance = Mock()
-            settings_instance.core.strict_envelope_mode = True
-            mock_settings.return_value = settings_instance
-
-            with patch(
-                "fapilog.core.worker.serialize_envelope",
-                side_effect=ValueError("Strict serialization error"),
-            ):
-                # In strict mode, should drop the entry (line 654: continue)
-                logger.info("Test message in strict mode")
-
-        await logger.stop_and_drain()
-
-        # In strict mode with serialization failure, behavior depends on implementation
-        # Just verify operation completed without crash
-        assert len(out) >= 0
-
-    @pytest.mark.asyncio
-    async def test_serialized_sink_exception_fallback(self) -> None:
-        """Test fallback when serialized sink raises exception."""
-        out: list[dict[str, Any]] = []
-
-        async def regular_sink(event: dict[str, Any]) -> None:
-            out.append(dict(event))
-
-        async def failing_serialized_sink(view: Any) -> None:
-            raise RuntimeError("Serialized sink error")
-
-        logger = SyncLoggerFacade(
-            name="serialized-sink-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=regular_sink,
-            sink_write_serialized=failing_serialized_sink,
-            serialize_in_flush=True,
-        )
-
-        logger.start()
-
-        # Should fall back to regular sink when serialized sink fails
-        logger.info("Test message with serialized sink failure")
-
-        await logger.stop_and_drain()
-
-        # Should fall back to regular sink path
-        assert len(out) >= 1
-
-
-class TestAsyncLoggerFacadeEdgeCases:
-    """Test AsyncLoggerFacade specific edge cases and exception paths."""
-
-    @pytest.mark.asyncio
-    async def test_async_logger_thread_mode_edge_cases(self) -> None:
-        """Test AsyncLoggerFacade in thread mode with edge cases."""
-        out: list[dict[str, Any]] = []
-
-        logger = AsyncLoggerFacade(
-            name="async-thread-edge-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=_create_async_sink(out),
-        )
-
-        # Test starting in thread mode when no event loop is running
-        # This should trigger thread mode initialization
-        logger.start()
-
-        await logger.info("Test message in thread mode")
-
-        await logger.stop_and_drain()
-
-        assert len(out) >= 1
-        assert len(out) >= 1
-
-    @pytest.mark.asyncio
-    async def test_async_logger_event_loop_mode_transitions(self) -> None:
-        """Test AsyncLoggerFacade event loop mode transitions."""
-        out: list[dict[str, Any]] = []
-
-        # Create logger in event loop context
-        logger = AsyncLoggerFacade(
-            name="async-event-loop-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=_create_async_sink(out),
-        )
-
-        # Should detect running event loop and use event loop mode
-        logger.start()
-
-        await logger.info("Test message in event loop mode")
-
-        # Test graceful shutdown
-        _ = await logger.stop_and_drain()
-
-        assert len(out) >= 1
-        assert len(out) >= 1
-
-
-class TestWorkerLifecycleExceptions:
-    """Test worker lifecycle exception handling."""
-
-    def test_worker_thread_exception_handling(self) -> None:
-        """Test exception handling in worker thread lifecycle."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="worker-exception-test",
-            queue_capacity=16,
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=1,
-            drop_on_full=False,
-            sink_write=lambda e: out.append(e),
-        )
-
-        # Mock threading.Event to cause issues in worker startup
-        original_event = threading.Event
-
-        def failing_event():
-            event = original_event()
-            # Make wait() fail after first call
-            original_wait = event.wait
-            call_count = 0
-
-            def failing_wait(timeout=None):
-                nonlocal call_count
-                call_count += 1
-                if call_count > 1:
-                    raise RuntimeError("Event wait failed")
-                return original_wait(timeout)
-
-            event.wait = failing_wait
-            return event
-
-        with patch("threading.Event", side_effect=failing_event):
-            logger.start()
-            logger.info("Test message with worker exception")
-
-            # Should handle worker exceptions gracefully
-            asyncio.run(logger.stop_and_drain())
-
-            # Message processing might be affected but shouldn't crash
-            assert len(out) >= 1
-
-
-class TestDiagnosticsExceptionPaths:
-    """Test diagnostics emission exception handling."""
-
-    def test_diagnostics_warn_exception_handling(self) -> None:
-        """Test exception handling when diagnostics.warn fails."""
-        out: list[dict[str, Any]] = []
-
-        logger = SyncLoggerFacade(
-            name="diagnostics-exception-test",
-            queue_capacity=2,  # Small queue to trigger backpressure
-            batch_max_size=4,
-            batch_timeout_seconds=0.01,
-            backpressure_wait_ms=0,  # No wait, immediate drop
-            drop_on_full=True,
-            sink_write=lambda e: out.append(e),
-        )
-
-        logger.start()
-
-        # Mock diagnostics.warn to raise an exception
-        with patch(
-            "fapilog.core.diagnostics.warn",
-            side_effect=RuntimeError("Diagnostics error"),
-        ):
-            # Fill the queue to trigger backpressure and diagnostics
-            for i in range(10):
-                logger.info(f"Message {i}")
-
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle diagnostics exceptions gracefully
-        # Some messages should be processed, some dropped
-        assert len(out) >= 1
-        assert logger._dropped >= 1
