--- tests/unit/test_core_logger_priority2_coverage.py
@@ -1,666 +1,0 @@
-"""
-Priority 2 Coverage Tests for Core Logger
-
-This module targets the remaining critical uncovered paths:
-1. Queue High Watermark Updates (Lines 683-685, 687-688)
-2. Serialization Fallback Paths (Lines 609-611, 620-622, 637-638)
-3. Metrics Exception Paths (Lines 709-710)
-
-These tests focus on monitoring, backpressure, and serialization robustness.
-"""
-
-import asyncio
-import os
-import threading
-import time
-from typing import Any
-
-import pytest
-
-from fapilog.core.logger import SyncLoggerFacade
-from fapilog.metrics.metrics import MetricsCollector
-
-
-def _create_async_sink(collected: list[dict[str, Any]]):
-    """Create an async sink function."""
-
-    async def async_sink(event: dict[str, Any]) -> None:
-        collected.append(dict(event))
-
-    return async_sink
-
-
-class TestQueueHighWatermarkUpdates:
-    """Test queue high watermark updates (Lines 683-685, 687-688)."""
-
-    def test_queue_high_watermark_updates_during_async_enqueue(self) -> None:
-        """Test queue high watermark updates during async enqueue (Line 683-685)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create metrics collector to track high watermark updates
-        metrics = MetricsCollector(enabled=True)
-
-        logger = SyncLoggerFacade(
-            name="high-watermark-test",
-            queue_capacity=16,
-            batch_max_size=8,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit many messages to create queue pressure and trigger high watermark
-        # updates
-        for i in range(50):
-            logger.info(f"watermark-test {i}")
-
-        # Allow processing time for high watermark updates
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 50
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert result.queue_depth_high_watermark > 0
-        if logger._worker_thread is not None:
-            assert not logger._worker_thread.is_alive()
-
-    def test_queue_high_watermark_updates_with_metrics_collection(self) -> None:
-        """Test queue high watermark updates with metrics collection (Line 687-688)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create metrics collector
-        metrics = MetricsCollector(enabled=True)
-
-        logger = SyncLoggerFacade(
-            name="metrics-watermark-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to create backpressure
-        for i in range(20):
-            logger.info(f"metrics-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify high watermark was tracked
-        assert result.queue_depth_high_watermark > 0
-        if logger._worker_thread is not None:
-            assert not logger._worker_thread.is_alive()
-
-    @pytest.mark.skipif(
-        os.getenv("CI") == "true", reason="Timing-sensitive; skipped in CI"
-    )
-    def test_queue_high_watermark_updates_with_backpressure_scenarios(self) -> None:
-        """Test queue high watermark updates with backpressure scenarios."""
-        collected: list[dict[str, Any]] = []
-
-        # Create metrics collector
-        metrics = MetricsCollector(enabled=True)
-
-        logger = SyncLoggerFacade(
-            name="backpressure-watermark-test",
-            queue_capacity=4,  # Small queue to create backpressure
-            batch_max_size=2,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages from multiple threads to create backpressure
-        def submit_messages(thread_id: int):
-            for i in range(10):
-                logger.info(f"thread-{thread_id}-{i}")
-                time.sleep(0.001)
-
-        threads = []
-        for i in range(3):
-            thread = threading.Thread(target=submit_messages, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        # Allow threads to submit messages
-        time.sleep(0.05)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Wait for threads to finish
-        for thread in threads:
-            thread.join()
-
-        # Verify high watermark was tracked during backpressure
-        assert result.queue_depth_high_watermark > 0
-        assert result.submitted >= 25
-        assert result.dropped >= 0
-        if logger._worker_thread is not None:
-            logger._worker_thread.join(timeout=0.5)
-            assert not logger._worker_thread.is_alive()
-
-
-class TestSerializationFallbackPaths:
-    """Test serialization fallback paths (Lines 609-611, 620-622, 637-638)."""
-
-    def test_serialization_envelope_exception_with_strict_mode(self) -> None:
-        """Test envelope serialization exception with strict mode (Line 609-611)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a sink that will receive the fallback
-        async def test_sink(event: dict[str, Any]) -> None:
-            collected.append(dict(event))
-
-        logger = SyncLoggerFacade(
-            name="strict-serialization-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=test_sink,
-            serialize_in_flush=True,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages
-        for i in range(3):
-            logger.info(f"strict-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 3
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_serialization_json_fallback_exception(self) -> None:
-        """Test JSON fallback serialization exception (Line 620-622)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a sink that will receive the fallback
-        async def test_sink(event: dict[str, Any]) -> None:
-            collected.append(dict(event))
-
-        logger = SyncLoggerFacade(
-            name="json-fallback-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=test_sink,
-            serialize_in_flush=True,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages
-        for i in range(3):
-            logger.info(f"json-fallback-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 3
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_serialized_sink_exception_fallback(self) -> None:
-        """Test serialized sink exception fallback (Line 637-638)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a sink that fails on serialized path but works on fallback
-        call_count = 0
-
-        async def failing_serialized_sink(event: dict[str, Any]) -> None:
-            nonlocal call_count
-            call_count += 1
-            if call_count % 2 == 0:  # Fail every other call
-                raise RuntimeError("Serialized sink failure")
-            collected.append(dict(event))
-
-        async def fallback_sink(event: dict[str, Any]) -> None:
-            collected.append(dict(event))
-
-        logger = SyncLoggerFacade(
-            name="serialized-fallback-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=fallback_sink,
-            sink_write_serialized=failing_serialized_sink,
-            serialize_in_flush=True,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages
-        for i in range(6):
-            logger.info(f"serialized-fallback-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 6
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_serialization_strict_mode_vs_best_effort(self) -> None:
-        """Test serialization strict mode vs best effort behavior."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a sink that will receive the fallback
-        async def test_sink(event: dict[str, Any]) -> None:
-            collected.append(dict(event))
-
-        logger = SyncLoggerFacade(
-            name="strict-vs-best-effort-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=test_sink,
-            serialize_in_flush=True,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages
-        for i in range(3):
-            logger.info(f"strict-vs-best-effort-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 3
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-
-class TestMetricsExceptionPaths:
-    """Test metrics exception paths (Lines 709-710)."""
-
-    def test_metrics_flush_latency_exception_handling(self) -> None:
-        """Test metrics flush latency exception handling (Line 709-710)."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that will fail on record_flush
-        class FailingMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def record_flush(self, batch_size: int, latency_seconds: float):
-                self.call_count += 1
-                if self.call_count % 2 == 0:  # Fail every other call
-                    raise RuntimeError("Metrics flush failure")
-                await super().record_flush(batch_size, latency_seconds)
-
-        metrics = FailingMetricsCollector()
-
-        logger = SyncLoggerFacade(
-            name="metrics-flush-exception-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to trigger flush with metrics
-        for i in range(8):
-            logger.info(f"metrics-flush-exception-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 8
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_metrics_sink_error_exception_handling(self) -> None:
-        """Test metrics sink error exception handling."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that will fail on record_sink_error
-        class FailingSinkErrorMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def record_sink_error(self, sink: str | None):
-                self.call_count += 1
-                if self.call_count % 2 == 0:  # Fail every other call
-                    raise RuntimeError("Metrics sink error failure")
-                await super().record_sink_error(sink)
-
-        metrics = FailingSinkErrorMetricsCollector()
-
-        # Create a sink that fails intermittently
-        call_count = 0
-
-        async def failing_sink(event: dict[str, Any]) -> None:
-            nonlocal call_count
-            call_count += 1
-            if call_count % 3 == 0:  # Fail every 3rd call
-                raise RuntimeError("Sink failure")
-            collected.append(dict(event))
-
-        logger = SyncLoggerFacade(
-            name="metrics-sink-error-exception-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=failing_sink,
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to trigger sink failures
-        for i in range(12):
-            logger.info(f"metrics-sink-error-exception-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 12
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_metrics_queue_high_watermark_exception_handling(self) -> None:
-        """Test metrics queue high watermark exception handling."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that will fail on set_queue_high_watermark
-        class FailingHighWatermarkMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def set_queue_high_watermark(self, value: int):
-                self.call_count += 1
-                if self.call_count % 2 == 0:  # Fail every other call
-                    raise RuntimeError("Metrics high watermark failure")
-                await super().set_queue_high_watermark(value)
-
-        metrics = FailingHighWatermarkMetricsCollector()
-
-        logger = SyncLoggerFacade(
-            name="metrics-high-watermark-exception-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to trigger high watermark updates
-        for i in range(16):
-            logger.info(f"metrics-high-watermark-exception-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 16
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-    def test_metrics_backpressure_wait_exception_handling(self) -> None:
-        """Test metrics backpressure wait exception handling."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that will fail on record_backpressure_wait
-        class FailingBackpressureMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def record_backpressure_wait(self, count: int):
-                self.call_count += 1
-                if self.call_count % 2 == 0:  # Fail every other call
-                    raise RuntimeError("Metrics backpressure wait failure")
-                await super().record_backpressure_wait(count)
-
-        metrics = FailingBackpressureMetricsCollector()
-
-        logger = SyncLoggerFacade(
-            name="metrics-backpressure-exception-test",
-            queue_capacity=2,  # Small queue to create backpressure
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to trigger backpressure
-        for i in range(10):
-            logger.info(f"metrics-backpressure-exception-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 10
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert logger._worker_thread is None
-
-
-class TestIntegrationScenarios:
-    """Test integration scenarios combining multiple Priority 2 areas."""
-
-    def test_high_watermark_with_serialization_fallback_and_metrics_exceptions(
-        self,
-    ) -> None:
-        """Test high watermark with serialization fallback and metrics exceptions."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that fails on multiple operations
-        class FailingIntegrationMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def set_queue_high_watermark(self, value: int):
-                self.call_count += 1
-                if self.call_count % 3 == 0:  # Fail every 3rd call
-                    raise RuntimeError("Integration high watermark failure")
-                await super().set_queue_high_watermark(value)
-
-            async def record_flush(self, batch_size: int, latency_seconds: float):
-                self.call_count += 1
-                if self.call_count % 4 == 0:  # Fail every 4th call
-                    raise RuntimeError("Integration flush failure")
-                await super().record_flush(batch_size, latency_seconds)
-
-        metrics = FailingIntegrationMetricsCollector()
-
-        logger = SyncLoggerFacade(
-            name="integration-test",
-            queue_capacity=8,
-            batch_max_size=4,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-            serialize_in_flush=True,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages to trigger all scenarios
-        for i in range(20):
-            logger.info(f"integration-test {i}")
-
-        # Allow processing time
-        time.sleep(0.1)
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify results
-        assert result.submitted >= 20
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert result.queue_depth_high_watermark > 0
-        if logger._worker_thread is not None:
-            logger._worker_thread.join(timeout=0.5)
-            assert not logger._worker_thread.is_alive()
-
-    @pytest.mark.skipif(
-        os.getenv("CI") == "true", reason="Timing-sensitive; skipped in CI"
-    )
-    def test_concurrent_high_watermark_and_metrics_exceptions(self) -> None:
-        """Test concurrent high watermark and metrics exceptions."""
-        collected: list[dict[str, Any]] = []
-
-        # Create a metrics collector that fails under pressure
-        class FailingConcurrentMetricsCollector(MetricsCollector):
-            def __init__(self):
-                super().__init__(enabled=True)
-                self.call_count = 0
-
-            async def set_queue_high_watermark(self, value: int):
-                self.call_count += 1
-                if self.call_count % 5 == 0:  # Fail every 5th call
-                    raise RuntimeError("Concurrent high watermark failure")
-                await super().set_queue_high_watermark(value)
-
-        metrics = FailingConcurrentMetricsCollector()
-
-        logger = SyncLoggerFacade(
-            name="concurrent-test",
-            queue_capacity=16,
-            batch_max_size=8,
-            batch_timeout_seconds=0.05,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=_create_async_sink(collected),
-            metrics=metrics,
-        )
-
-        # Start in thread mode
-        logger.start()
-
-        # Submit messages from multiple threads
-        def submit_messages(thread_id: int):
-            for i in range(15):
-                logger.info(f"concurrent-thread-{thread_id}-{i}")
-                time.sleep(0.0001)  # Faster submission
-
-        threads = []
-        for i in range(4):
-            thread = threading.Thread(target=submit_messages, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        # Allow threads to submit messages
-        time.sleep(0.1)  # More time for submission
-
-        # Test drain
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Wait for threads to finish
-        for thread in threads:
-            thread.join()
-
-        # Verify results - be more flexible with submission count
-        assert result.submitted >= 40  # Allow for some messages to be dropped
-        assert result.processed >= 0
-        assert result.dropped >= 0
-        assert result.queue_depth_high_watermark > 0
-        if logger._worker_thread is not None:
-            logger._worker_thread.join(timeout=0.5)
-            assert not logger._worker_thread.is_alive()
