--- tests/unit/test_core_logger_priority4_coverage.py
@@ -1,840 +1,0 @@
-"""
-Priority 4 Coverage Tests for Core Logger
-
-This module targets the remaining critical uncovered paths:
-1. Advanced Error Handling (Lines 176-177, 285-286, 386-387, 427-428)
-2. Complex Exception Scenarios (Lines 496-497, 507-508, 521-522)
-3. Advanced Threading Edge Cases (Lines 580-591)
-4. Complex Serialization Paths (Lines 609-611, 620-622, 637-638, 650-651)
-5. Advanced Queue Management (Lines 683-685, 747-760, 880, 882, 902, 926-927)
-
-These tests focus on enterprise robustness, edge cases, and complex error scenarios.
-"""
-
-import asyncio
-import time
-from typing import Any
-from unittest.mock import Mock, patch
-
-from fapilog.core.logger import SyncLoggerFacade
-
-
-class TestAdvancedErrorHandling:
-    """Test advanced error handling scenarios."""
-
-    def test_drain_with_runtime_error_during_loop_check(self) -> None:
-        """Test drain when RuntimeError occurs during loop check (Lines 176-177)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit some messages
-        for i in range(5):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test the drain logic - this should exercise lines 176-177
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Verify proper cleanup
-        assert logger._worker_thread is None
-        assert logger._worker_loop is None
-        assert result.submitted >= 5
-        assert result.processed >= 0
-
-    def test_error_dedupe_with_diagnostics_exception(self) -> None:
-        """Test error dedupe when diagnostics.warn raises exception (Lines 285-286)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Mock diagnostics.warn to raise exception
-        with patch(
-            "fapilog.core.diagnostics.warn", side_effect=RuntimeError("Warn failed")
-        ):
-            logger.error("duplicate error message")
-            logger.error("duplicate error message")  # Should be suppressed
-            logger.error("duplicate error message")  # Should be suppressed
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should still process messages even if diagnostics fails
-        assert len(collected) >= 1
-
-    def test_cross_thread_enqueue_with_loop_exception(self) -> None:
-        """Test cross-thread enqueue when loop operations fail (Lines 386-387)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=5,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Fill the queue to trigger cross-thread path
-        for i in range(10):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle the scenario gracefully
-        assert result.submitted >= 10
-
-    def test_cross_thread_warn_exception_handling(self) -> None:
-        """Test cross-thread warning when diagnostics.warn fails (Lines 427-428)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=5,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Fill the queue to trigger cross-thread path
-        for i in range(10):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle the scenario gracefully
-        assert result.submitted >= 10
-
-
-class TestComplexExceptionScenarios:
-    """Test complex exception handling scenarios."""
-
-    def test_exception_serialization_with_complex_exc_info(self) -> None:
-        """Test exception serialization with complex exc_info (Lines 496-497)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Create a complex exception scenario
-        try:
-            raise ValueError("Test exception")
-        except ValueError:
-            logger.error("Complex exception", exc_info=True)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle complex exc_info
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-    def test_exception_serialization_with_custom_exc_info(self) -> None:
-        """Test exception serialization with custom exc_info tuple (Lines 507-508)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Create custom exc_info
-        custom_exc_info = (ValueError, ValueError("Custom"), None)
-        logger.error("Custom exception", exc_info=custom_exc_info)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle custom exc_info
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-    def test_exception_serialization_with_exc_parameter(self) -> None:
-        """Test exception serialization with exc parameter (Lines 521-522)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Create exception object
-        exc = RuntimeError("Runtime exception")
-        logger.error("Runtime exception", exc=exc)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle exc parameter
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-
-class TestAdvancedThreadingEdgeCases:
-    """Test advanced threading edge cases."""
-
-    def test_worker_thread_cleanup_with_exception_during_join(self) -> None:
-        """Test worker thread cleanup when join raises exception (Lines 580-591)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit some messages
-        for i in range(5):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test drain - this should exercise the worker cleanup logic
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle cleanup gracefully
-        assert result.submitted >= 5
-
-    def test_worker_loop_cleanup_with_exception(self) -> None:
-        """Test worker loop cleanup when operations fail (Lines 580-591)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit some messages
-        for i in range(5):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test drain - this should exercise the loop cleanup logic
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle cleanup gracefully
-        assert result.submitted >= 5
-
-
-class TestComplexSerializationPaths:
-    """Test complex serialization paths."""
-
-    def test_serialization_fallback_with_errors_serialize_exception_failure(
-        self,
-    ) -> None:
-        """Test serialization fallback when serialize_exception fails (Lines 609-611)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Mock serialize_exception to raise exception
-        with patch(
-            "fapilog.core.errors.serialize_exception",
-            side_effect=RuntimeError("Serialize failed"),
-        ):
-            logger.error("Exception message", exc=ValueError("Test"))
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle serialization failure gracefully
-        assert len(collected) >= 1
-
-    def test_serialization_fallback_with_exception_during_metadata_update(
-        self,
-    ) -> None:
-        """Test serialization fallback when metadata update fails (Lines 620-622)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Test exception serialization - this should exercise the metadata update path
-        logger.error("Exception message", exc=ValueError("Test"))
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle metadata update gracefully
-        assert len(collected) >= 1
-
-    def test_serialization_fallback_with_exception_during_exc_info_normalization(
-        self,
-    ) -> None:
-        """Test serialization fallback when exc_info normalization fails (Lines 637-638)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Mock getattr to raise exception during exc_info normalization
-        with patch("builtins.getattr", side_effect=RuntimeError("Getattr failed")):
-            logger.error("Exception message", exc=ValueError("Test"))
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle exc_info normalization failure gracefully
-        assert len(collected) >= 1
-
-    def test_serialization_fallback_with_exception_during_sys_exc_info(
-        self,
-    ) -> None:
-        """Test serialization fallback when sys.exc_info fails (Lines 650-651)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Mock sys.exc_info to raise exception
-        with patch("sys.exc_info", side_effect=RuntimeError("Exc info failed")):
-            logger.error("Exception message", exc_info=True)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle sys.exc_info failure gracefully
-        assert len(collected) >= 1
-
-
-class TestAdvancedQueueManagement:
-    """Test advanced queue management scenarios."""
-
-    def test_queue_high_watermark_update_with_large_qsize(self) -> None:
-        """Test queue high watermark update with large queue size (Lines 683-685)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=100,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit many messages rapidly to trigger high watermark updates
-        for i in range(150):
-            logger.info(f"message {i}")
-
-        time.sleep(0.05)
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should update high watermark during processing
-        assert result.queue_depth_high_watermark > 0
-
-    def test_queue_high_watermark_persistence_across_drains(self) -> None:
-        """Test queue high watermark persistence across multiple drains (Lines 687-688)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=50,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # First burst to establish high watermark
-        for i in range(60):
-            logger.info(f"burst1 message {i}")
-
-        time.sleep(0.05)
-        result1 = asyncio.run(logger.stop_and_drain())
-        first_hwm = result1.queue_depth_high_watermark
-
-        # Start a new logger for the second burst
-        logger2 = SyncLoggerFacade(
-            name="test2",
-            queue_capacity=50,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger2.start()
-
-        # Second burst to potentially increase high watermark
-        for i in range(60):
-            logger2.info(f"burst2 message {i}")
-
-        time.sleep(0.05)
-        result2 = asyncio.run(logger2.stop_and_drain())
-        second_hwm = result2.queue_depth_high_watermark
-
-        # High watermark should be reasonable
-        assert first_hwm > 0
-        assert second_hwm > 0
-
-    def test_advanced_drain_scenarios_with_complex_worker_state(self) -> None:
-        """Test advanced drain scenarios with complex worker state (Lines 747-760)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit messages and trigger complex worker state
-        for i in range(30):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test drain - this should exercise complex worker state handling
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle complex worker state gracefully
-        assert result.submitted >= 30
-
-    def test_drain_with_none_worker_thread_and_loop(self) -> None:
-        """Test drain when worker thread and loop are None (Line 762)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=10,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit some messages
-        for i in range(5):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test normal drain behavior - this should exercise the drain logic
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle drain gracefully
-        assert result.submitted >= 5
-
-    def test_queue_operations_with_edge_case_capacities(self) -> None:
-        """Test queue operations with edge case capacities (Lines 880, 882, 902, 926-927)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=1,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Test with minimal capacity to trigger edge cases
-        logger.info("first message")
-        logger.info("second message")  # Should be dropped
-
-        time.sleep(0.05)
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle edge case capacity gracefully
-        assert result.submitted >= 1
-        assert result.dropped >= 1
-
-
-class TestComplexIntegrationScenarios:
-    """Test complex integration scenarios."""
-
-    def test_complex_error_deduplication_with_window_rollover(self) -> None:
-        """Test complex error deduplication with window rollover (Lines 996-1001)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Mock settings for very short dedup window
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            settings_instance = Mock()
-            settings_instance.core.error_dedupe_window_seconds = 0.01  # 10ms window
-            mock_settings.return_value = settings_instance
-
-            # Submit duplicate errors with timing to trigger window rollover
-            logger.error("duplicate error")
-            time.sleep(0.005)  # Within window
-            logger.error("duplicate error")  # Should be suppressed
-            time.sleep(0.02)  # Outside window - rollover
-            logger.error("duplicate error")  # New window
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle complex deduplication scenarios
-        # Note: deduplication is working (see stdout), but only 1 message is collected
-        # This is expected behavior - duplicate messages are suppressed
-        assert len(collected) >= 1
-
-    def test_complex_error_deduplication_with_summary_emission(self) -> None:
-        """Test complex error deduplication with summary emission (Lines 1018-1040)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=30,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Mock settings and diagnostics
-        with patch("fapilog.core.settings.Settings") as mock_settings:
-            settings_instance = Mock()
-            settings_instance.core.error_dedupe_window_seconds = 0.01
-            mock_settings.return_value = settings_instance
-
-            with patch("fapilog.core.diagnostics.warn"):
-                # Submit many duplicate errors to trigger summary
-                for _i in range(50):
-                    logger.error("duplicate error")
-                    time.sleep(0.001)  # Small delay
-
-                # Wait for window rollover
-                time.sleep(0.02)
-                logger.error("duplicate error")  # Should trigger summary
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should emit summary and handle complex deduplication
-        assert len(collected) >= 2
-
-    def test_exception_serialization_with_complex_traceback(self) -> None:
-        """Test exception serialization with complex traceback (Lines 1051-1052)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Create complex exception with traceback
-        def nested_function():
-            def inner_function():
-                raise ValueError("Nested exception")
-
-            inner_function()
-
-        try:
-            nested_function()
-        except ValueError:
-            logger.error("Complex nested exception", exc_info=True)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle complex traceback serialization
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-    def test_exception_serialization_with_custom_exception_attributes(self) -> None:
-        """Test exception serialization with custom exception attributes (Lines 1065, 1075)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-        )
-        logger.start()
-
-        # Create custom exception with attributes
-        class CustomException(Exception):
-            def __init__(self, message, custom_attr):
-                super().__init__(message)
-                self.custom_attr = custom_attr
-
-        exc = CustomException("Custom message", "custom_value")
-        logger.error("Custom exception", exc=exc)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should handle custom exception attributes
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-    def test_exception_serialization_with_frame_limiting(self) -> None:
-        """Test exception serialization with frame limiting (Lines 1086-1087)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-            exceptions_enabled=True,
-            exceptions_max_frames=5,
-        )
-        logger.start()
-
-        # Create deep call stack
-        def deep_function(depth):
-            if depth <= 0:
-                raise ValueError("Deep exception")
-            return deep_function(depth - 1)
-
-        try:
-            deep_function(20)  # Deep call stack
-        except ValueError:
-            logger.error("Deep exception", exc_info=True)
-
-        time.sleep(0.05)
-        asyncio.run(logger.stop_and_drain())
-
-        # Should limit frames according to configuration
-        assert len(collected) >= 1
-        event = collected[0]
-        metadata = event.get("metadata", {})
-        assert "error.type" in metadata or "error.message" in metadata
-
-
-class TestFinalCleanupAndShutdown:
-    """Test final cleanup and shutdown scenarios."""
-
-    def test_complex_shutdown_with_multiple_worker_states(self) -> None:
-        """Test complex shutdown with multiple worker states (Lines 1384-1395)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=20,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit messages and create complex worker state
-        for i in range(15):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown - this should exercise complex shutdown scenarios
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle complex shutdown gracefully
-        assert result.submitted >= 15
-
-    def test_shutdown_with_exception_during_worker_cleanup(self) -> None:
-        """Test shutdown when worker cleanup raises exception (Lines 1406-1407)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=15,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit messages
-        for i in range(10):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown - this should exercise worker cleanup scenarios
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle cleanup gracefully
-        assert result.submitted >= 10
-
-    def test_shutdown_with_exception_during_loop_cleanup(self) -> None:
-        """Test shutdown when loop cleanup raises exception (Lines 1416-1417)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=15,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit messages
-        for i in range(10):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown - this should exercise loop cleanup scenarios
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle cleanup gracefully
-        assert result.submitted >= 10
-
-    def test_final_cleanup_with_resource_cleanup_exception(self) -> None:
-        """Test final cleanup when resource cleanup raises exception (Lines 1455, 1460-1462, 1466, 1469)."""
-        collected: list[dict[str, Any]] = []
-        logger = SyncLoggerFacade(
-            name="test",
-            queue_capacity=15,
-            batch_max_size=4,
-            batch_timeout_seconds=0.1,
-            backpressure_wait_ms=1,
-            drop_on_full=True,
-            sink_write=lambda e: collected.append(e),
-        )
-        logger.start()
-
-        # Submit messages
-        for i in range(10):
-            logger.info(f"message {i}")
-
-        # Allow processing time
-        time.sleep(0.05)
-
-        # Test shutdown - this should exercise resource cleanup scenarios
-        result = asyncio.run(logger.stop_and_drain())
-
-        # Should handle resource cleanup gracefully
-        assert result.submitted >= 10
